requirements.txt
Flask
Werkzeug
DeepFace
python-dotenv
requests
tensorflow